Sorting of Files before PDF creation.
loop_files_for_item() was looping through the Files in DESC order by
File->ID. This was causing multipage PDF creation with the last page of
the document first and then descending to the first page. Added sorting
in so that multiple page PDF files are now created in order sorted on
Original Filename.